Itinerary header wip
